+++ HISTORY.rst
@@ -4,6 +4,7 @@
 
 Unreleased Changes
 ------------------
+* Data Stack files (``*.invest.json``, ``*.invest.tar.gz``) can now be dragged and dropped on an InVEST model window, which will prompt the UI to load that parameter set.
 * Spatial inputs to Coastal Blue Carbon are now aligned as part of the model.
   This resolves a longstanding issue with the model where inputs would need to
   perfectly overlap (even down to pixel indices), or else the model would yield